Feature/volume slider

+++ includes/GUI/GuiElement.h
@@ -8,11 +8,11 @@ protected:
     Vector2 position{};
 
     virtual Vector2 setElementSize(unsigned x, unsigned y) = 0;
-    virtual Vector2 setElementPosition(float x, float y) = 0;
+    virtual Vector2 setElementPosition(unsigned x, unsigned y) = 0;
 
 public:
     GuiElement(const unsigned screenWidth, const unsigned screenHeight) {
-        windowSize = { static_cast<float>(screenWidth), static_cast<float>(screenHeight) };
+        windowSize = {static_cast<float>(screenWidth), static_cast<float>(screenHeight) };
     }
 
     virtual ~GuiElement() = default;
